lifeline-backend-master/requirements.txt
aioredis==1.2.0
amqp==2.4.1
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
Babel==2.6.0
billiard==3.5.0.5
boto3==1.9.9
botocore==1.12.93
celery==4.2.1
certifi==2018.11.29
channels==2.1.6
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.2
dj-database-url==0.5.0
dj-email-url==0.1.0
Django==1.11.15
django-admin-view-permission==1.8
django-autoslug==1.9.3
django-cleanup==2.1.0
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-extensions==2.1.2
django-filter==2.0.0
django-material==1.0.0
django-modeltranslation==0.12.2
django-redis==4.9.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
flower==0.9.2
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jmespath==0.9.3
kombu==4.3.0
Markdown==2.6.11
MarkupSafe==1.1.0
msgpack==0.5.6
openapi-codec==1.3.2
openpyxl==2.5.5
Pillow==5.3.0
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
PyJWT==1.7.1
PyDocX==0.9.10
python-dateutil==2.8.0
pytz==2018.9
redis==2.10.6
requests==2.21.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
tornado==5.1.1
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
zope.interface==4.6.0
xhtml2pdf==0.2.3
hl7==0.3.4
django-ses==0.8.10
pydocx==0.9.10

